A9.8 Progress, Refactored Cli.define_main_menu
Changed name to _define_menus to indicate it isn't normally accessed.
Used only as a constructer now, no setter style "If not None:" code
Adjusted tests to show that the constructor resets  values of the attrs
Reverted prior change which was making them class attrs because the
attrs were not easily accessable as class attrs, and make more sense as
instance attrs anyways.
49 Tests. 100% Coverage. 111 Statements.